[Editorial] Remove reference to portrait in preface
The print edition includes a portrait of Daisy Ashford as a child. I've removed the phrase "which is here reproduced as frontispiece to her sublime work" since the portrait is absent in this edition. However, I've kept the rest of the paragraph, as the reader can assume Barrie himself had a portrait of the writer in front of him.