Revert most instances of "Mr." to "Mr"
Because the page scans sometimes have the period and sometimes do not, it seems the distinction could be purposeful. In the spirit of keeping the author's original spelling etc., I'm deleting the period where it is also absent in the page scans (i.e. undoing some of what typogrify did automatically)